+++ nix/patches/default.nix
@@ -31,4 +31,7 @@
   # https://github.com/nh2/static-haskell-nix/pull/91
   static-haskell-nix-postgrest-openssl-linking-fix =
     ./static-haskell-nix-postgrest-openssl-linking-fix.patch;
+
+  static-haskell-nix-hasql-notifications-openssl-linking-fix =
+    ./static-haskell-nix-hasql-notifications-openssl-linking-fix.patch;
 }