+++ nix/static-haskell-package.nix
@@ -18,6 +18,7 @@ let
       static-haskell-nix
       [
         patches.static-haskell-nix-postgrest-openssl-linking-fix
+        patches.static-haskell-nix-hasql-notifications-openssl-linking-fix
       ];
 
   patchedNixpkgs =